+++ lib/mix/lib/mix/tasks/loadpaths.ex
@@ -43,9 +43,9 @@ defmodule Mix.Tasks.Loadpaths do
       case Version.parse_requirement(req) do
         {:ok, req} ->
           unless Version.match?(System.version, req) do
-            Mix.raise Mix.ElixirVersionError, target: config[:app] || Mix.Project.get,
-                                              expected: req,
-                                              actual: System.version
+            raise Mix.ElixirVersionError, target: config[:app] || Mix.Project.get,
+                                          expected: req,
+                                          actual: System.version
           end
         :error ->
           Mix.raise "Invalid Elixir version requirement #{req} in mix.exs file"